AirSwipe Crash Prevention
Added a protective measure to check the amount of entities that are
about to have their velocities changed before moving them all. If there
are too many entities around it will damage them without changing their
velocity, that way we don't lag out and crash due to the firing of too
many EntityCollisionEvents.