StockAbility Checker and small clean up
Check if there is already a stockability enabled or if there is one when
someone makes a new (fan made) ability. this allows you to create
ability with same name as default ones as long as you turn off the
default one. Also cleaned up some small parts, and will clean up the
command class in feature