FallingSand config!
Owners can now decide if bending affects fallingsand, there are 2 types
of fallingsand:
TNT and fallingsand, for both there are 2 config options if bending
should affect it, and if it affects it, than what should be the strength
multiplier be? 1.0 is the default strength and 0.1 is 1/10 of the normal
strength.

For developers:
new Methods, Methods.setVelocity(entity, vector)