Added multiple methods that allow for removing ability instances
AirShield,FireShield: isWithinShield() checks if a location is within
the shield.
AirSpout,WaterSpout: Updated removeSpouts to return true if an instance
was removed.
AirSwipe: Added a method that removes AirSwipes around a location within
a radius.

Added a Method.blockAbilities that takes a player, list of ability
names, location, and a radius. Using the ability names, it tries to
remove any instances of each of the abilities at a specific location
using the radius. If AirShield or FireShield was included as an ability
they will not be removed, instead they will just return true.

Added a Methods.isWithinShields that checks if a location is within
either a FireShield or an AirShield.