Particle update
The big particle update :D
Updated particle library, can be found here:
http://bukkit.org/threads/1-8-particleeffect-v1-6.154406/

Also made a new method to if region is protected from build, this one is
less CPU/database intensive! Since it will not ask for permissions but
will check the region, which always be local on the RAM/flatfile,
instead of being in a permissions database/flatfile which is always
slower!