Optimized Bloodbending a bit
If you were to bloodbend an entity initially close to you, the distance
between you and that entity while bloodbending would be the same. This
made throwing difficult as if you were to hit the entity (because it's
within arm range) the move wouldn't work. I set the distance value to 6,
so that when you Bloodbend an entity the distance between it and
yourself is always 6 unless you point at a wall or the floor.