The floood is real!!! :D Fix!
WaterArms is not set up properly in build 1 of PK we changed all water
and stationair water to (BYTE) 8 so nothing flood and this one is again
0. It floods everything :/ anyways here the fix :)

Also for PEX & GM & ZPerms fix (for as far as I know)
WG asks the permissions plugin to check AGAIN for a permissions line, if
one of these plugins is using a MySQL it is a party, any move will lag
the server enormous. This way WG just checks if there is a flag set, IF
a player has the permissions line from WG to build anywhere it is still
ignored, so basically the same, only faster and 1000 times faster when
using MySQL perms.