Configuration and bPlayer Changes
added isBender(bPlayer); to bendingplayer. It returns true if the player can bend any of the elements and it returns false if the player can't bend any element.
Changed how custom elements are made, added String color and String prefix, so a developer can set a custom color and a custom prefix to his element.
Changed how SubElements are made, added String color, so the developer can set a custom color. But beware, the developer should only choose one color. If he chooses 2 the color of the 1st declared SubElement will be chosen.