Fixed EarthTunnel bug (#860)
## Fixes
* EarthTunnel bug
    > * https://trello.com/c/O1zvQT0N/831-earthtunnel-cant-be-used-on-gravel-sand-red-sand-blocks-that-are-fell
This bug was caused by blocks with gravity (sand, gravel, etc) falling into already existing TempBlocks and attempting to use EarthTunnel on these blocks. Previously it would result in nothing happening, and the block would be stuck.
This issue was caused by TempBlock not correctly checking the material type (and data!) of the already existing TempBlock when attempting to create a new instance on a pre-existing TempBlock.
This was fixed by instead of checking if `newtype` and `newdata` on the old and new TempBlock are different, it checks the actual type of the block by using `temp.block.getType()`, to get the actual values. This is because the already created TempBlock has the same type as what is trying to be set (air), and so the code to set the type again never is called.