Update GeneralMethods.java - Fix
On line 145, NETHER_WARTS_BLOCK is used in the list of non-opaque blocks. This seems to be a mistake with the material ID- it is the nether wart block added in 1.10; the material that should most likely be used here is NETHER_STALKS.

I may also create a different pull request in this area which may improve the CPU usage for checking these materials significantly (right now, the plugin loops through the whole array every time it checks if it's opaque or not opaque).