Update AirSwipe.java - Fix
There is an incorrectly written check for whether or not the block is breakable. The move is trying to see if the individual block is a torch made by Illumination, but this will always return false with how the code is written. Illumination.getBlocks() returns a ConcurrentHashMap of key and value type <TempBlock,Player>... this code is trying to see if the map contains a block as a key (this has keys that are TempBlocks, but not Blocks).

This fixes this problem using the method provided by Illumination.

##Fix
(second part of issue)
- Issue: https://github.com/ProjectKorra/ProjectKorra/issues/882