Change Illumination.java
Changed Illumination#isIlluminationTorch(block) to be more efficient, and deprecated Illumination#getBlocks() in favor of using isIlluminationTorch().

getBlocks() has been used in a less efficient manner as a way of determining if the block is created by Illumination, or even used incorrectly, by not getting the TempBlock of a block before checking if it contains it. This could possibly be used to remove torches created by illumination, but if the tempblock isn't reverted after being removed, it might not be for a while.