Update TempBlock.java - Additions
Added ability and preventPhysics parameters to TempBlock. This is currently not a requirement for making a tempblock; the old constructors will still work perfectly fine. The methods for retrieving these I have also added @Deprecated to, because they are currently not used in the code anywhere (these should be removed when they are used thoroughly, ideally in place of abilities' getBlocks().contains(tempblock).

Both of these are really needed because too many abilities have their own methods of telling you if the block is created by it and if physics should be prevented; there is no universal ability.ownsTempBlock(TempBlock) method. With this fully put to use throughout the code, PKListener and add-on abilities can easily specify if a block should or shouldn't be affected by physics, and it will be much more efficient and easier to see which ability made the tempblock.