Added database cooldowns (#893)
**Overview**
In order for cooldowns to be saved through server restarts, a new database table has been introduced (pk_cooldowns) which will hold information for specific player's who have abilities on cooldown. This data will be saved onDisable and loaded back in onEnable. This is currently being used for the `/pk choose` command and the `AvatarState` ability.

**Utilizing the API**
To create a cooldown for a player which will be saved in this database table, simply provide an extra boolean argument to the end of BendingPlayer#addCooldown(...) defining this feature. A couple new BendingPlayer#addCooldown(...) methods have been introduced to take this extra argument, leaving it false for all the previous addCooldown methods.