Changed WaterArms displayBoundMsg condition
Under onPlayerSlotChange, it would send the player messages with info on their subabilities if BendingPreview was set to false. Now, the condition has changed, so that if "DisplayBoundMsg" (a new config option under WaterArms) is set to true, it would do the same thing. This allows for more configurability instead of tying this to another config option. Sometimes people want BendingPreview off but don't want to see the WaterArms messages.