Allowed FastSwim with WaterArms
This is a very debatable change, but I believe WaterArms is overall better this way. Of course it might not make as much sense this way because we never saw this happen in the shows, but for whatever reason WaterArms grapple is much easier when FastSwim is permitted. I've been using WaterArms for years, and when you could use FastSwim with it, Grapple was more consistent than ever, and it was a great experience to use WaterArms. Now, Grapple is harder to use; it's spotty and inconsistent. Even though it's probably a bug, FastSwim changes this.