Added a new FastSwim config option
"AllowWaterArms" is now a configurable boolean, with default set to false. If false, FastSwim will be prohibited while WaterArms is being used; if true, FastSwim will be allowed. FastSwim with WaterArms somehow makes Grapple much easier to use, and I think people would greatly appreciate this option, even if it is sort of a bug. Generally more options for users and more configurability is better.