- Found out, that FireKick's issue was not because of origin's height, but because of finite vectors. So, returned origin of FireKick back, how it was, added variable 'origin'.
-Fixed GeneralMethods#getOrthogonalVector's issue:
when it was trying to get orthogonal vector for a vector with 0 x and z, the vector became zero-vector. After normalize() this vector became NaN NaN NaN, so FireKick's FireComboStreams didn't work because of checkFinite thrown Exception.

Now, if you use this method like it was, then you'll get randomly rotated Orthogonal vector, if x and z were 0.
Now you'd be better to add yaw to this method, or use Location with direction instead of Vector.