BlockSource.getWaterSourceBlock refactor.
Fixes the bug, when you disabled dynamic sources in config, then you could use IceBlast from water, leaving a hole in water surface.

I didn't really get the logic of original 'allowWaterBottles' check, so I reversed it.

If water bottles aren't allowed, then it'll ignore gotten above sourceBlock, if distance to it less or equal 3. (because it could be a block, created with waterbottle, and they are not allowed, so I don't want to get it as a source?)

Actually I even tried to comment whole this allowWaterBottles check section, and it didn't affect on bottlebending at all, lol.

Also I removed !ElementalAbility.isAir check in bottom expression, because it's unnecessary, because you are checking later, that block is water, ice, plant, snow or cauldon for sure, so it can't be air anyway.