the command /bending add <element> checks to see if the command sender has permission to add specific elements.
the command checks for the permission bending.command.add.<element>, and prevents the command sender from adding the element to the target if the permission is denied.

this also applies to the command /bending add all, and individually checks if the command sender has permission to add each element individually.