Optimized TempBlocks
Reduced the lags caused from TempBlocks reverting. The main cause was due to removing blocks going through the entire queue to find the block to remove, when it could have just been removed from the top.

Additionally, there is no longer checks to see if it is already in the queue in some instances. This also caused a lot of pointless CPU processing