Various bug fixes
Fixes WhoCommand lag when a player is offline doing look ups on the main thread. Fixes a TempArmor bug where you can swap your Armor by right clicking with an armor pieces in your hand and it will swap whatever temparmor you're wearing with it. It will give you the temparmor as a piece and then delete the actual armor you swapped with when it gets reverted