Fix AirShield range underflow and Catapult/EarthPillar disabled bug (#1394)
## Fixes
* Fixes
    > * https://github.com/ProjectKorra/ProjectKorra/issues/1357
    > * https://github.com/ProjectKorra/ProjectKorra/issues/1354

## First issue:
If a player would receive FallDamage while having the Earth element, the DensityShift passive would enable if falling on Earth-Bendable block. But if `EarthPillars` is disabled and `Catapult` is the hovered slot, the plugin tries to instantiate a `EarthPillars` instance which isn't possible due to it being disabled and it throws an exception in the event which leads to fall damage.

### How I fixed it
Simply check if EarthPillars is enabled before trying to instantiate it.


## Second issue
If AirShield maxRadius setting would be 1, it would result in an infinite loop.

### How I fixed it
Fixed the issue by making the `di` Integer never be able to be 0 to let the for loop increment